requirements.txt
scrapy
flask
shelve
python-scrapyd-api
sevice-identity
pyopenssl >= 0.15